Initial import.
These sources are converted from the old Z80asm format (only changes to DEFM text string syntax), and
with slight modifications to some of the text strings.
This code is under development for the new flash eprom library, and is not API nor feature sync'ed yet
with the latest changes
of the library.