Final implementation of polling for formatable file areas in available slots. Also, new message if there's no formatable file areas available, both when FlashStore boots and when user select the format command.
Still needs to be implemented (format cmd prepared for it):
A new pop-up where the user may select a particular card if there are
multiple choices; the window will display potential file space for new file
areas, and if a re-format is possible, display the current no. active files
for that slot. The window will only allow formatable slots to be selected.

A formatable slot is defined as a Flash Card that has erase/write support
for the particular slot and that space is available on the card for a file area.