NULL file now automatically save to file area as first file, if a file area is formatted on an Intel card. This function was previously an integrated part of FlashStore. The NULL file handling should however be part of the standard library.
The reason for the NULL file:
---------------------------------------------------------------------------------------------
Due to a strange side effect with Intel Flash Chips, a special "NULL" file
is saved as the first file to the bottom of the Card. These bytes occupies the
first bytes that otherwise could be interpreted as a random boot command for
the Intel chip - the behaviour is an Intel chip suddenly gone into command
mode for no particular reason.
The NULL file prevents this behaviour by saving a file that avoids any
kind of boot commands which sends the chip into command mode when the
card has been inserted into a Z88 slot.
---------------------------------------------------------------------------------------------