battlow library routine removed from library.
This routine was a implementation mistake based on wrong interpretation of the hardware section of the Developers Notes. The battlow interrupt is not a static on/off switch that indicates a battery low status or not, but only a single interrupt event triggered by the Blink, that is acknowledged by thew system and therefore cleared.

There is no well-defined battery low state constant signal to be read from the hardware. It is only displayed statically on the screen.