New optimisation feature: Only execute the Card ID polling routine for INTEL or AMD Flash on RAM System Stack if the executing library is executing on the same slot that is going to be polled for a Card / Manufacturer ID.
Example: If the FlashEprCardId library is going to poll slot 3 for a Card ID, and is running in slot 2, then just directly execute the polling code located on the card. If the library is located in slot 2 and slot 2 is going to polled for a Card ID, then relocate the polling subroutine of the library to the RAM System Stack and execute it there.