New optimisation feature: Only execute the write block routine for INTEL or AMD Flash on RAM System Stack if the executing library is executing on the same slot (when the chip is in command mode during write cycles it doesn't function as a normal read array mode memory device).
Example: If the FlashEprWriteBlock library is going to write a block of data to slot 3 and is running in slot 2, then the code will just directly execute the polling code located on the card in slot 2. If the library and data to be written is both located in slot 2, then the core write routine is relocated to the RAM System Stack and executed there.