Several major changes: 1) API change; define Flash chip type in A (or 0 to indicate dynamic programming of both INTEL & AMD chip). If 0 is supplied, then the actual chip type is returned in A when the library returns succesfully.    This optimises the internal static polling for flash type for every data block write. If a file is being written, the first block write will use A = 0 (get chip type before data block programming), and the    following calls to this library already have the chip type established (which was returned after first call to FlashEprWriteBlock).
2) Avoid NMI interrupts in executing flash chip routine by temporarily blocking all interrupts from the Blink (replaces the old DI/EI functionality).

3)New optimisation feature:
Only execute the write block routine for INTEL or AMD Flash on RAM System Stack if the executing library is executing on the same slot (when the chip is in command mode during write cycles it doesn't function as a normal read array mode memory device).
Example: If the FlashEprWriteBlock library is going to write a block to slot 3 and is running in slot 2, then the code will just directly execute the write block routine located on the card in slot 2. If the library and data to be written is both located in slot 2, then the core write routine is relocated to the RAM System Stack and executed there.