Implemented BBC BASIC CRC32 routine before calling RomUpdate application at $25C0.
The CLI (text) version is the text presentation of the pure BBC BASIC boot program that performs the CRC32 check. The CLI text file is executed on the Z88 and a BAS version is saved back to the compile project where it is binary included into the assembler project.

The BBC BASIC binary program only needs to be updated if the CLI version is changed.

Currently, the added boot.bas is fully functional; it opens the 'romupdate.cfg' file and reads the CRC32 value defined for the 'romupdate.bas' program just loaded into memory. A complete CRC32 is performed of the loaded program and compared to the value from the config file. If they match (the binary program was successfully loaded to the Z88 without being corrupted), the actual RomUpdate native assembler application is executed.