Initial draft implementation of Z88 Font BitMap source code generator.
The tool is used to scan a binary copy of the lores1 and hires1 bitmaps that are located in the existing Z88 ROMs. The tool is supplied a filename of the extracted font binary (copied from bank 7, offset 0).

TODO: 
Generate font bitmaps without the bit7,6 hidden token table
Generate auto-generated source code of the hidden token table and implement it
as a normal token-table.