Mpm V1.1 b7 (22/11/2005): New feature: Generate Global Definitions file without full linking (excluding expression evaluation).
When two independent project binaries needs to cross reference address information in the other project binary, and both projects are needs the 'other' projects address information to fully compile, we have mutual address link problem.

The trick is to resolve one set of XDEF addresses in the first project without full linking (and XREF address evaluation). The generated DEFC file will be the basis for the second project binary that may perform a full link (including address relocation evaluation) and also generate it's own Global Definitions constant file, which in turn is used for a second full compile & link of the first project binary.