Migrated the RomUpdate CRC value away from 'romupdate.cfg' file and into it's own file, 'romupdate.crc'.
This allows the romupdate.bas & romupdate.crc to be kept separate from the actual apps & related configuration files to be updated. This enables Rakewell on their website to have a central place for the RomUpdate utility (the latest & greatest version) and keep all the Zips with updateable/addable apps in another location.