OZ ROM: Renamed BANK_MTH to OZBANK_MTH compile variable to be more consistent with naming scheme of other bank variable definitions.
Relocated MTH static structures to bank 8, and introduced new file area sub-watermark in Application ROM header which identifies a file area that is available from the top of the OZ rom and downwards. The file area is defined in 64K boundaries to reflect Flash chip architecture and to avoid the file area colliding with OZ banks (when re-formatting file area).