Sysvar reorganized in 6 parts. Kernel variables (some space has been freed for growing lowram) Preserved area (96 bytes protected during soft reset) Hardware area (where the keyboard buffer have been inclued) Structures area (buffers and handles) External area (for not kernel related system variables Index application, GN calls) Process area (used for each application instance containing the mailbox and the application stack)
Reset routine have been modified to correctly erase all the bank (except the preserved area).

Lots of space is free. It let us a lot of freedom for future implementations. Interruption or serial code could certainly be transfered in RAM.

Note that $1000-$17FF is 2K dedicated to swap and is only used on unexpanded machines. This large space (in a such critical bank) could  be used for other purposes.