/z88apps/RomUpdate: When updating an OZ Rom in slot X: Adjust the destination bank number (of the bank file to be blown) to the OZ.x slot number. This makes it possible to alter the romupdate.cfg file in the "OZ.x" to the slot number of the destination.
Originally "OZ.0" or "OZ.1" were designed as the type OZ to be blown. OZ for slot 0 is still different than for slot 1; altering an OZ.0 to slot 1 would not give a problem (since there is no "ZOZ" header" - which will ignore the manipulated OZ in slot 1 designed for slot 0). Changing an OZ.1 to OZ.2 or OZ.3 would not create a problem either, since the final updated 1Mb card is going to be moved into slot 1 (followed by a HARD RESET) to get working.