OZ/Intuition integration: First (early) work of Intuition V1.2.dev version - this release recognizes OZ 4.2 API and is called through RST 08H instruction. For now, the existing OZ compile scripts automatically integrate Intuition as part of the OZ ROM, for development purposes. OZ ROM Release compile scripts will not contain Intuition. The runtime area is located in the OZ system variables; Intuition will be transparent for OZ. Further testing needs to be done to see if Intuition can trace OZ completely. OZ code that uses upper 8K of segment 0 will be re-written to reserve this segment for Intuition.
Intuition for segment 0/OZ has undergone major code optimization to re-integrate the .VA command (view stack). Intuition is activated through the RST 08H instruction and will be dynamically bound into upper 8K of segment 0. When the debugger is released, the original bank binding of upper segment 0 will be restored (this will be handled in LOWRAM area, part of OZ).

TO DO: The RST 20H execution interface will be changed to automatically trace OZ calls, and not release the debugger temporarily. Also, tracing of OS_POUT will currently soft reset Z88, when Intuition executes this OZ V4.2 system call. The plan is also to activate Intuition through the OS_IN/TIN interface (special keypress) that returns to the caller with Fc=1, A=RC_SUSP. This allows to get control of the Z80 processor, when OZ is waiting for a keypress. It will be interesting to see how this solution is implemented.