/z88apps/installer: Fixed issue where Index invoked Bootstrap on card insertion/removal
Whenever Index rebuilds the application list (ie on a card insertion/
removal) it invokes the first application marked with "AT_Boot" in the DOR.
This meant that Bootstrap would be invoked at such times; it would then
detect (via the presence of a lock on the booted.yes file) that this was not
a reset and immediately exit. This could be annoying; the usual sequence of
events was:

1. Insert or remove a card whilst in Index.
2. Enter an application.
3. Press Index; at this point, Bootstrap would be briefly invoked and then
   quit, leaving you back in the application that you were using in step 2.
4. Press Index; this time, you'd get to Index correctly.

The fix employed here is to extend the substitution for the OS_NQ call where
BC=NQ_Ain. This now always substitutes such calls (not just if IX=Diary handle)
and clears the AT_Boot bit in A before returning to the caller. This prevents
Index from seeing Bootstrap as a bootable application when it rebuilds the
application list. Bootstrap is still invoked after a reset (soft or hard),
because at this point the substitutions have not been installed (Bootstrap
does this itself).

The booted.yes file is now redundant, so has been removed.