/tools/makeapp: Added "generateCardId" directive to automatically generate card Ids
It was noticed that out of a total of 20 cards (including compilations) that can be
built from the z88apps directory, only 7 different card Ids were being used. Since
card Id management is a bit of a chore, it's obviously often overlooked.

This checkin adds a new directive "generateCardId". If present in a loadmap,
the card Id (at $3ff8-$3ff9 in bank $3f) will be automatically generated using
14 bits from the CRC32 of the card images (bits 7 & 15 must always be set to zero).