/tools/forth: Improved workaround for FlashStore CLI issue
FlashStore appears to consume input from the CLI even when busy.
This causes the boot.cli file to continue before all files have been
restored from/saved to the card. To workaround this a delay (.D 3000)
was introduced. However, this is not sufficient to build WhatNow?
An improved workaround seems to be to add multiple delays.