/oz, /tools/makeapp, OZ-10: Reinstate revision numbering system
Reinstated automatic build revision numbering system, which was removed when
the project migrated from svn to git.

The new system uses git describe and produces version information of the
form "TTTTTTT-RRRR-gCCCCCC" where TTTTTTT represents the last tag (usually
of the form OZ_Vx.y), RRRR is the number of commits since the tag, and CCCCCCC
is the SHA1 of the most recent commit. If the tag name is unexpectedly
long, it will be truncated at the left.