EAZ-22 When a file has been transferred to desktop, the hosting panel does not reflect the file size EAZ-30 file in subdir on z88 doesn't copy to Dir on Desktop
Added a relative path entry to each selected item in the z88.
This way, the relative path can be appended to the destination Desktop path.

Also, created a work around to a QT bug, where the size of a newly transfered file isnt reflected.
The work around was to rename the file to filename.xfer, then renaming it back.