/tools/makeapp, OZ-15: Fix build number on help screen if latest tag is not for OZ
If the latest tag was for a different component, the build number for OZ
would use that tag, resulting in an incorrect help screen display.

The latest tag containing the pattern "OZ_V" is now used instead.

Note that automatic build revision text could potentially be added to other
components in a similar fashion, by adding a makeapp configuration entry
specifying the tag pattern to be matched.