EAZ-18 When EazyLink is started the first time, implement a wizard that tries to find the connected Z88 EAZ-88 Drag and Drop using Windows Explorer fails EAZ-78 Imp/Exp transfers fail
Added a Setup Wizard that runs the first time the user runs Eazylink2.  It allows user to choose
mode of operation. ie Between ezylink mode, or Imp-Export mode.

The Windows Drag and drop issue should be resolved. I was using the QT::Separator option to compare for windows
paths, but, QT in wondows alos uses the '/' separator.

The Addition of the Setup Wizard should fix the EAZ-78 issue, where the default startup used to try to find the Z88
using the Eazylink Protocol.