/z88apps/epfetch: Fixed identification of RAM/ROM hybrid cards
File areas on RAM/ROM hybrid cards were not detected correctly as
it was assumed that the file area started at the bottom bank of the
card. In hybrid cards, the bottom portion is all RAM, only the upper
banks are ROM.

Fixed by calculating the actual file area base bank using the card
size in its header.