/z88apps/ziputils/unzip, ZUZ-2: WA for failure to properly close partially-written files
If a no room error is encountered whilst writing a file, the output handle for
the partially-written file was closed before the input handle for the same
file. In theory this should be fine, but the first GN_CL returns an RC_ERR
error, and the partially-written file is listed in the filer with no size.

Reversing the order of the closes allows both GN_CL calls to succeed, and OZ
automatically deletes the partially-written file.

Fixes ZUZ-2